requirements.txt
asgiref==3.6.0
Django==4.2
djangorestframework==3.14.0
postgres==4.0
psycopg2-binary==2.9.6
psycopg2-pool==1.1
pytz==2023.3
sqlparse==0.4.3
tzdata==2023.3

python-dateutil~=2.8.2